@@ -1,5 +1,5 @@
 [tool.black]
-target-version = ["py36"]
+target-version = ["py37"]
 
 # black compatible isort
 [tool.isort]
